create, edit and delete work as expected, just when a user is logged in.  pending fixing view profile after show 1 post
