+++ src/main/java/com/codeup/springbootblog/controllers/PostsController.java
@@ -55,45 +55,8 @@ public class PostsController {
     }
 
 
-    @GetMapping("posts/new")
-    // to catch the form
-    // "/new" is to create a new post and "/create" is the one that is going to show that it was created.
-    public String showCreatePostForm(Model viewModel){
-        Post post = new Post();
-        viewModel.addAttribute("post", post);
-        return "/blog_template/new";
-    }
-
-    @PostMapping("posts/create")
-    // we are not using pathvariable here because is comming from a form
-    public String savePost(@ModelAttribute Post post){
-        // this is coming form the db, and UserWithRoles is not mapped.
-        User user = (User) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
-        post.setUser(usersDao.findOne(user.getId()));
-        postService.save(post);
-        return "redirect:/";
-    }
-
-    @GetMapping("posts/{id}/edit")
-    public String showEditForm(@PathVariable long id, Model viewModel){
-        Post post = postService.findOne(id);
-        postService.save(post);
-        viewModel.addAttribute("post", post);
-        return "/blog_template/edit";
-    }
-
-    @GetMapping("posts/{id}/delete")
-    public String deletePost(@PathVariable long id, Model viewModel){
-        postService.delete(id);
-        return "redirect:/";
-    }
 
 
-    @PostMapping("posts/edit")
-    public String updatePost(@ModelAttribute Post post){
-        postService.update(post);
-        return "redirect:/";
-    }
