+++ src/main/resources/templates/blog_template/new.html
@@ -37,7 +37,7 @@
                             <!--<h4 class="modal-title" id="myModalLabel">Please register!!!</h4>-->
                         </div>
                         <div class="modal-body">
-                            <form th:action="@{/posts/create}" method="post" th:object="${post}">
+                            <form th:action="@{/user/posts/create}" method="post" th:object="${post}">
                                 <div class="row">
                                     <div class="col-lg-3">
                                         <h6>New post title:</h6>
