+++ src/main/resources/templates/fragments/blog_template_navbar.html
@@ -22,7 +22,7 @@
                     </li>
 
                     <li class="nav-item">
-                        <a class="nav-link" href="/posts/new"><span class="glyphicon glyphicon-tag"></span> Create A New Post</a>
+                        <a class="nav-link" href="/user/posts/new"><span class="glyphicon glyphicon-tag"></span> Create A New Post</a>
                     </li>
 
                     <li class="nav-item" sec:authorize="isAuthenticated()">
